Move resources package to shared namespace
