+++ src/main/java/com/amannmalik/mcp/content/ContentBlock.java
@@ -3,8 +3,8 @@ package com.amannmalik.mcp.content;
 import com.amannmalik.mcp.annotations.Annotations;
 import com.amannmalik.mcp.client.sampling.MessageContent;
 import com.amannmalik.mcp.prompts.PromptContent;
-import com.amannmalik.mcp.server.resources.Resource;
-import com.amannmalik.mcp.server.resources.ResourceBlock;
+import com.amannmalik.mcp.resources.Resource;
+import com.amannmalik.mcp.resources.ResourceBlock;
 import com.amannmalik.mcp.validation.InputSanitizer;
 import com.amannmalik.mcp.validation.MetaValidator;
 import jakarta.json.JsonObject;
