+++ src/main/java/com/amannmalik/mcp/resources/ListResourceTemplatesRequest.java
@@ -1,4 +1,4 @@
-package com.amannmalik.mcp.server.resources;
+package com.amannmalik.mcp.resources;
 
 import com.amannmalik.mcp.validation.MetaValidator;
 import jakarta.json.JsonObject;
