+++ src/main/java/com/amannmalik/mcp/resources/ListResourcesResult.java
@@ -1,4 +1,4 @@
-package com.amannmalik.mcp.server.resources;
+package com.amannmalik.mcp.resources;
 
 import com.amannmalik.mcp.util.Immutable;
 import com.amannmalik.mcp.validation.MetaValidator;
