+++ src/main/java/com/amannmalik/mcp/resources/ResourceSubscription.java
@@ -1,4 +1,4 @@
-package com.amannmalik.mcp.server.resources;
+package com.amannmalik.mcp.resources;
 
 public interface ResourceSubscription extends AutoCloseable {
     @Override
