+++ src/main/java/com/amannmalik/mcp/resources/ResourceTemplate.java
@@ -1,29 +1,25 @@
-package com.amannmalik.mcp.server.resources;
+package com.amannmalik.mcp.resources;
 
 import com.amannmalik.mcp.annotations.Annotations;
 import com.amannmalik.mcp.util.DisplayNameProvider;
 import com.amannmalik.mcp.validation.*;
 import jakarta.json.JsonObject;
 
-public record Resource(
-        String uri,
+public record ResourceTemplate(
+        String uriTemplate,
         String name,
         String title,
         String description,
         String mimeType,
-        Long size,
         Annotations annotations,
         JsonObject _meta
 ) implements DisplayNameProvider {
-    public Resource {
-        uri = UriValidator.requireAbsolute(uri);
+    public ResourceTemplate {
+        uriTemplate = UriTemplateValidator.requireAbsoluteTemplate(uriTemplate);
         name = InputSanitizer.requireClean(name);
         title = InputSanitizer.cleanNullable(title);
         description = InputSanitizer.cleanNullable(description);
         mimeType = InputSanitizer.cleanNullable(mimeType);
-        if (size != null && size < 0) {
-            throw new IllegalArgumentException("size must be >= 0");
-        }
         annotations = annotations == null ? Annotations.EMPTY : annotations;
         MetaValidator.requireValid(_meta);
     }
