+++ src/main/java/com/amannmalik/mcp/resources/UnsubscribeRequest.java
@@ -1,11 +1,11 @@
-package com.amannmalik.mcp.server.resources;
+package com.amannmalik.mcp.resources;
 
 import com.amannmalik.mcp.validation.MetaValidator;
 import com.amannmalik.mcp.validation.UriValidator;
 import jakarta.json.JsonObject;
 
-public record ReadResourceRequest(String uri, JsonObject _meta) {
-    public ReadResourceRequest {
+public record UnsubscribeRequest(String uri, JsonObject _meta) {
+    public UnsubscribeRequest {
         uri = UriValidator.requireAbsolute(uri);
         MetaValidator.requireValid(_meta);
     }
